+++ admin/inc/theme_functions.php
@@ -164,20 +164,40 @@ function get_page_subtitle($echo = true) {
  * Get Page Summary
  *
  * Retrieves the summary of a page, optionally echoing it.
- * Allows the option to strip HTML tags from the summary.
  *
  * @since 2025.2
  * @uses $summary
  * @uses strip_decode()
  *
  * @param bool $echo Optional, default is true. If true, the summary is echoed.
- * @param bool $strip_tags Optional, default is true. If true, HTML tags are stripped from the summary.
  * @return string|null The summary of the page, either echoed or returned based on the $echo parameter.
  */
-function get_page_summary($echo = true, $strip_tags = true) {
+function get_page_summary($echo = true) {
 	global $summary;
 	$result = strip_decode((string) $summary);
-	if ($strip_tags) $result = strip_tags($result);
+	if ($echo) {
+		echo $result;
+	} else {
+		return $result;
+	}
+}
+
+/**
+ * Get clean page summary
+ *
+ * Retrieves the summary of a page, stripped of html tags and decoded, optionally echoing it.
+ *
+ * @since 2025.2
+ * @uses $summary
+ * @uses strip_decode()
+ * @uses strip_tags()
+ *
+ * @param bool $echo Optional, default is true. If true, the summary is echoed.
+ * @return string|null The clean summary of the page, either echoed or returned based on the $echo parameter.
+ */
+function get_page_clean_summary($echo = true) {
+	global $summary;
+	$result = stip_tags(strip_decode((string) $summary));
 	if ($echo) {
 		echo $result;
 	} else {