+++ CHANGELOG.md
@@ -15,7 +15,7 @@ All notable changes of GetSimple Legacy CMS will be documented in this file.
 - Added the field *Page Subtitle* to store subtitle or long title of the page.
 - Added theme function `get_page_subtitle()` to echo or return the page subtitle.
 - Added the field *Page Summary* to store summary of the page.
-- Added theme function `get_page_summary()` to echo or return the page summary.
+- Added theme functions `get_page_summary()` and `get_page_clean_summary()` to echo or return the page summary or the page summary stripped of html tags.
 - Added the field *Page Featured Image* to store URL of featured image of the page.
 - Added theme function `get_page_featured_image()` to echo or return the URL of the page featured image.
 - Added theme function `get_component_title()` to echo or return title of a component.